+++ src/main/java/com/airwallex/operator/UndoOperator.java
@@ -3,6 +3,7 @@ package com.airwallex.operator;
 import com.airwallex.lex.TokenMatcher;
 import com.airwallex.mode.UndoMode;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class UndoOperator extends Operator {
@@ -23,7 +24,7 @@ public class UndoOperator extends Operator {
         return null;
     }
 
-    public Double executeInternal(List<NumberToken> numberTokenList) {
+    public BigDecimal executeInternal(List<NumberToken> numberTokenList) {
         if (this.getNumberStack() != null) {
             if (UndoMode.LexicalMode.equals(numberTokenList.get(0).getUndoMode())) {
                 //simply remove it, as we already pop up the fields,nothing need to do here
